+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 6.0.0-alpha.6 (DD.MM.YYYY)
 
+### Breaking Changes
+
+-   **terra-checkbox** removed getter and setter for `value` and `valueChange` emitter. Use `ngModel` and `ngModelChange` instead.
+-   **terra-slider** remove deprecated fields `inputValue` and `inputValueChange`.
+
 ### Features
 
 -   **dependencies** updated all `jasmine` and `karma` packages to the latest compatible versions.